+++ app/javascript/src/components/Dashboard/Notes/constants.js
@@ -1,54 +1,13 @@
 import { t } from "i18next";
-import * as yup from "yup";
 
 const AVATAR_IMAGE_URL = "https://randomuser.me/api/portraits/women/90.jpg";
 
-export const NOTES_FORM_INITIAL_FORM_VALUES = {
+export const INITIAL_FORM_VALUES = {
   title: "",
   description: "",
   role: "",
   tags: [],
 };
-
-export const NOTES_FORM_VALIDATION_SCHEMA = yup.object({
-  title: yup.string().required(t("validation.title")),
-  description: yup.string().required(t("validation.description")),
-  role: yup.object().required(t("validation.role")),
-  tags: yup.array().min(1, t("validation.tag")).required(t("validation.tag")),
-});
-
-export const TAGS = [
-  {
-    label: "Getting Started",
-    value: "getting_started",
-  },
-  {
-    label: "Onboarding",
-    value: "Onboarding",
-  },
-  {
-    label: "User Flow",
-    value: "user_flow",
-  },
-  {
-    label: "UX",
-    value: "ux",
-  },
-  {
-    label: "Bugs",
-    value: "bugs",
-  },
-  {
-    label: "V2",
-    value: "v2",
-  },
-];
-
-export const ROLES = [
-  { label: "Oliver Smith", value: "523621a4-ac39-48e5-8d86-eba9310d2055" },
-  { label: "Eve Smith", value: "523621a4-ac39-48e5-8d86-eba9310d5098" },
-  { label: "Sam Smith", value: "523621a4-ac39-48e5-8d86-eba9310d5685" },
-];
 
 export const SEGMENT_BLOCKS = [
   { count: 80, label: t("notes.segments.europe") },
