+++ app/javascript/src/translations/en.json
@@ -26,7 +26,6 @@
     "addNote": "Add Note",
     "addNewNote": "Add New Note",
     "searchNotes": "Search Name, Email, Phone Number, Etc.",
-    "newNotePaneTitle": "Title",
     "segments": {
       "europe": "Europe",
       "middleEast": "Middle-East",
